fixed build and overrides
