+++ src/lib/crypto/kimchi_bindings/js/dune
@@ -17,7 +17,8 @@
    bindings/verifier-index.js
    native/native-overrides/oracles.js
    native/native-overrides/util.js
-   native/native-overrides/circuit.js))
+   native/native-overrides/circuit.js
+   native/native-overrides/pasta-plonk-proof.js))
  (instrumentation
   (backend bisect_ppx))
  (preprocess
