+++ src/lib/crypto/kimchi_bindings/js/native/native_backend.js
@@ -1,4 +1,2 @@
 // Provides: plonk_napi
-var plonk_napi = require('@o1labs/o1js-darwin-arm64');
-
-module.exports = plonk_napi;
+var plonk_napi = require('@o1js/native-' + process.platform + '-' + process.arch)
