+++ src/lib/crypto/kimchi_bindings/js/node_js/node_backend.js
@@ -1,29 +1,18 @@
 // Provides: plonk_wasm
-var plonk_wasm = require('./plonk_wasm.js');
-var native = null;
-try {
-  native = require(`@o1labs/o1js-${process.platform}-${process.arch}`);
-} catch (e) {
-  // native not available, keep WASM
-}
+var plonk_wasm = (function() {
+  var wasm = require('./plonk_wasm.js');
 
-function snakeToCamel(name) {
-  return name.replace(/_([a-z])/g, function (_match, ch) {
-    return ch.toUpperCase();
-  });
-}
+  try {
+    var native =  require('@o1js/native-' + process.platform + '-' + process.arch)
 
-function override(functionName) {
-  if (!native) return;
-  var camel = snakeToCamel(functionName);
-  var impl = native[functionName] || native[camel];
-  if (typeof impl === 'function') {
-    plonk_wasm[functionName] = impl;
+    wasm["caml_pasta_fp_poseidon_block_cipher"] = native["caml_pasta_fp_poseidon_block_cipher"]
+    wasm["caml_pasta_fq_poseidon_block_cipher"] = native["caml_pasta_fq_poseidon_block_cipher"]
+  } catch (e) {
+    console.error(e)
+    console.log("native didn't load")
+    process.exit(1);
   }
-}
 
-// Overwrite only the functions that are already available in native
-override('caml_pasta_fp_poseidon_block_cipher');
-override('caml_pasta_fq_poseidon_block_cipher');
-override('prover_to_json');
-override('prover_index_from_bytes');
+  return wasm
+})()
+
