+++ src/lib/crypto/kimchi_bindings/js/web/web_backend.js
@@ -1,2 +1,5 @@
 // Provides: plonk_wasm
 var plonk_wasm = globalThis.plonk_wasm;
+
+// Provides: plonk_napi
+var plonk_napi = globalThis.plonk_wasm
